[MCUX-86415][LPI2C] Fixed MISRA C-2012 violations issues
- Rule 10.3 (fsl_lpi2c.c:936): add explicit (uint16_t) cast on the
  result of MIN() when assigning to the uint16_t tmpChunk variable to
  resolve narrowing essential-type violation
- Rule 10.8 (fsl_lpi2c.c:937, 1264): cast tmpChunk to uint32_t before
  subtracting 1U so the composite expression passed to LPI2C_MTDR_DATA()
  is not widened inside the macro, resolving the composite-expression
  cast-to-wider-type violation
- Rule 10.6 (fsl_lpi2c_edma.c:453, 655): cast 1U to uint32_t before
  the left-shift so the composite expression does not widen implicitly
  when assigned to the uint32_t maxTxFifo variable

+++ drivers/lpi2c/fsl_lpi2c.c
@@ -933,8 +933,8 @@ status_t LPI2C_MasterReceive(LPI2C_Type *base, void *rxBuff, size_t rxSize)
                 return result;
             }
 
-            uint16_t tmpChunk = MIN(rxSize - chunkSize, LPI2C_MAX_RX_SIZE);
-            base->MTDR = (uint32_t)kRxDataCmd | LPI2C_MTDR_DATA(tmpChunk - 1U);
+            uint16_t tmpChunk = (uint16_t)MIN(rxSize - chunkSize, LPI2C_MAX_RX_SIZE);
+            base->MTDR = (uint32_t)kRxDataCmd | LPI2C_MTDR_DATA((uint32_t)tmpChunk - 1U);
             chunkSize += tmpChunk;
         }
 
@@ -1261,7 +1261,7 @@ static void LPI2C_TransferStateMachineReadCommand(LPI2C_Type *base,
     stateParams->txCount--;
 
     uint16_t tmpChunk = MIN(handle->remainingBytes - handle->chunkSize, LPI2C_MAX_RX_SIZE);
-    base->MTDR = (uint32_t)kRxDataCmd | LPI2C_MTDR_DATA(tmpChunk - 1U);
+    base->MTDR = (uint32_t)kRxDataCmd | LPI2C_MTDR_DATA((uint32_t)tmpChunk - 1U);
     handle->chunkSize += tmpChunk;
 
     /* Move to transfer state. */
